==> notebooks/02_primary.ipynb <==
library(RPostgreSQL)
library(twang)
library(Matching)
library(tidyverse)

data_dir <- file.path("..", "data")
sql_dir <- file.path("..", "sql")

drv <- dbDriver("PostgreSQL")
con <- dbConnect(drv, dbname = "mimic")
con <- dbConnect(drv, dbname = "mimiciii", user = "postgres", password = "123")
dbSendQuery(con, "set search_path=echo,public,mimiciii;")

full_data <- dbGetQuery(con, "select * from merged_data")
head(full_data)

dbDisconnect(con)
dbUnloadDriver(drv)

to_factor <- function(x) {
    if(length(na.omit(unique(x))) <= 1) return(factor(x, levels = c(0, 1)))
    return(factor(x))
}

factor_vars <- full_data %>%
    names %>%
    grep("flag|abnormal|icd|sedative", ., value = TRUE) %>%
    c("gender", "first_careunit", "echo", "vent", "vaso",
      "icu_adm_hour", "mort_28_day")
factor_vars

weekday <- c("sun", "mon", "tues", "wednes", "thurs", "fri", "satur") %>%
paste0("day")
weekday

full_data <- full_data %>%
    mutate(echo_int = as.integer(echo)) %>%
    mutate(icu_adm_weekday = weekday[icu_adm_weekday + 1]) %>%
    mutate(mort_28_day_int = as.integer(mort_28_day)) %>%
    mutate_at(factor_vars, to_factor) %>%
    mutate(icu_adm_weekday = factor(icu_adm_weekday, levels = weekday)) %>%
    mutate(gender = relevel(gender, "M")) %>%
    mutate(first_careunit = relevel(first_careunit, "SICU"))
full_data %>% pull(echo) %>% head





full_data <- full_data %>%
  mutate(echo_int = as.integer(echo)) %>%
  mutate(icu_adm_weekday = as.numeric(icu_adm_weekday) + 1) %>%
  mutate(mort_28_day_int = as.integer(mort_28_day)) %>%
  mutate_at(factor_vars, to_factor) %>%
  mutate(icu_adm_weekday = factor(icu_adm_weekday, levels = weekday)) %>%
  mutate(gender = relevel(gender, "M")) %>%
  mutate(first_careunit = relevel(first_careunit, "SICU"))

feature_names <- full_data %>%
    names %>%
    keep(grepl("vs|lab|icd|age|gender|weight|saps|sofa|elix_score|vent|vaso|icu_adm|careunit|sedative", .)) %>%
    discard(grepl("vs|lab", .) & grepl("flag", .) & !grepl("bnp|troponin|kinase|cvp", .)) %>%
    discard(grepl("bnp|troponin|kinase|cvp", .) & !grepl("flag", .)) %>%
    discard(grepl("min|max", .)) %>%
    discard(grepl("abnormal", .))
feature_names
length(feature_names)

feature_names %>%
data.frame(feature = ., stringsAsFactors = FALSE) %>%
data.table::fwrite(file = file.path(data_dir, "C:/Users/yanghui/Desktop/features.csv"))


feature_names %>%
  data.frame(feature = ., stringsAsFactors = FALSE) %>%
  data.table::fwrite(file = "C:/Users/yanghui/Desktop/features.csv")

features <- full_data %>%
    select(!!!rlang::syms(feature_names))
head(features)






label_name <- "echo"

label <- full_data %>% pull(echo)
str(label)

fml <- feature_names %>%
    c("echo", .) %>%
    paste(collapse = " + ") %>%
    sprintf("mort_28_day ~ %s", .)
fml

unweighted <- glm(as.formula(fml), data = full_data, family = binomial, na.action = na.exclude)
summary(unweighted)
exp(cbind(OR = coef(unweighted), confint(unweighted)))



saveRDS(unweighted, file = file.path(data_dir, "multivariate_model.rds"))

fml <- feature_names %>%
    paste(collapse = " + ") %>%
    sprintf("echo_int ~ %s", .)
fml

echo_ps_ate <- ps(as.formula(fml),
                  data = full_data,
                  interaction.depth = 2,
                  shrinkage = 0.01,
                  perm.test.iters = 0,
                  estimand = "ATE",
                  verbose = FALSE,
                  stop.method = c("es.mean", "es.max", "ks.mean", "ks.max"),
                  n.trees = 10000,
                  train.fraction = 0.8,
                  cv.folds = 3,
                  n.cores = 8)

pred <- echo_ps_ate$ps$es.mean.ATE
full_data <- full_data %>% mutate(ps = pred)
ROCR::performance(ROCR::prediction(pred, label), "auc")@y.values %>% first

ft_importance <- summary(echo_ps_ate$gbm.obj,
                         n.trees = echo_ps_ate$desc$es.mean.ATE$n.trees,
                         plot = FALSE)

full_data <- full_data %>%
    mutate(ps_weight = get.weights(echo_ps_ate, stop.method = "es.mean"))

saveRDS(full_data, file = file.path(data_dir, "full_data_ps.rds"))



data_dir <- file.path("..", "data")
sql_dir <- file.path("..", "sql")

saveRDS(ft_importance, file = file.path(data_dir, "feature_importance.rds"))

primary_ipw <- glm(mort_28_day ~ echo, data = full_data,
                   weights = full_data$ps_weight, family = binomial)
summary(primary_ipw)
exp(cbind(OR = coef(primary_ipw), confint(primary_ipw)))

saveRDS(primary_ipw, file = file.path(data_dir, "ipw_model.rds"))

ipw_svydesign <- svydesign(ids = ~ icustay_id, weights = ~ ps_weight, data = full_data)

saveRDS(ipw_svydesign, file = file.path(data_dir, "ipw_svydesign.rds"))

fml <- feature_names %>%
    c(label_name, .) %>%
    paste(collapse = " + ") %>%
    sprintf("mort_28_day ~ %s", .)
fml

logi <- svyglm(as.formula(fml),
               family = quasibinomial,
               design = ipw_svydesign)

summary(logi)

exp(cbind(OR = coef(logi), confint(logi)))

saveRDS(logi, file = file.path(data_dir, "doubly_robust_all_model.rds"))
