==> notebooks/05_doubly_robust_unbalanced_covariates.ipynb <==
library(survey)
library(tidyverse)

data_dir <- file.path("..", "data")

full_data <- readRDS(file.path(data_dir, "full_data_ps.rds"))
head(full_data)

wtd_pval <- data.table::fread(file.path(data_dir, "wtd_pval.csv"), data.table = FALSE)
head(wtd_pval)

wpv <- wtd_pval %>%
    filter(pval < 0.05) %>%
    pull(cov)

wpv

ipw_svydesign <- readRDS(file.path(data_dir, "ipw_svydesign.rds"))

fml <- wpv %>%
c("echo", .) %>%
paste(collapse = " + ") %>%
sprintf("mort_28_day ~ %s", .)

fml

logi <- svyglm(as.formula(fml), family = quasibinomial, design = ipw_svydesign)

summary(logi)

exp(cbind(OR = coef(logi), confint(logi)))
